Typo in path in generate-webkitversion.pl https://bugs.webkit.org/show_bug.cgi?id=90883
Patch by Simon Hausmann <simon.hausmann@nokia.com> on 2012-07-12
Reviewed by Mark Rowe.

Correct default path to version config file. In practice it wasn't a problem because only the Qt
build system seems to use the script right now and passes the path to the version file as a
command line parameter.

* scripts/generate-webkitversion.pl:

Identifier: 108902@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@122427 268f45cc-cd09-0410-ab3c-d52691b4dbfc