[Qt][WK2] ASSERT: "!m_viewportItem->isMoving()" in QtViewportHandler::flickMoveEnded() https://bugs.webkit.org/show_bug.cgi?id=90875
Reviewed by Kenneth Rohde Christiansen.

Since MultiPointTouchArea and PinchArea use the childMouseEventFilter
method to filter touch events and because Flickable filters child mouse
events the canvas calls this function before propagating the touch event
to the WebView. Since Flickable does not accept touch events the canvas
tries to propagate a synthesized mouse event through the base class
childMouseEventFilter function which is accepted by Flickable and
interferes with the input events we send to Flicakble hence messes up
the internal state of the WebView.
This patch reimplements the virtual childMouseEventFilter method so that all
the mouse and touch events can be processed by WebKit before they arrive to
Flickable.

* UIProcess/API/qt/qquickwebview.cpp:
(QQuickWebView::childMouseEventFilter):
* UIProcess/API/qt/qquickwebview_p.h:

Identifier: 108913@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@122438 268f45cc-cd09-0410-ab3c-d52691b4dbfc