storage tests are flaky (crashing) on windows https://bugs.webkit.org/show_bug.cgi?id=90469
Patch by James Weatherall <wez@chromium.org> on 2012-07-12
Reviewed by Kentaro Hara.

Add a missing check that the underlying V8 object reference in a V8 NPObject is valid, and zero the NPObject's rootObject member when disposing it, to ensure that it won't be mistakenly touched after that point.

This patch is intended to resolve flakiness in the storage tests including:
  storage/indexeddb/mozilla/indexes.html
  storage/indexeddb/mozilla/key-requirements-inline-and-passed.html
  storage/websql/multiple-databases-garbage-collection.html

* bindings/v8/NPV8Object.cpp:
(WebCore::disposeUnderlyingV8Object):
Zero the NPObject's underlying rootObject.
(_NPN_EvaluateHelper):
Add check that the underlying V8 object reference is valid.

Identifier: 108962@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@122488 268f45cc-cd09-0410-ab3c-d52691b4dbfc