Teach CodeGenerator to support for static, readonly, attributes https://bugs.webkit.org/show_bug.cgi?id=88920 <rdar://problem/11650330>
Reviewed by Oliver Hunt.

Update the parser to be able to accept the static keyword for attribute. We will treat static attributes
like custom static functions. They call the implementing class directly, and pass in the ExecState as a script context.

* bindings/scripts/CodeGeneratorJS.pm:
(GetAttributeGetterName): Factor out the construction of the attribute getter function name.
(GetAttributeSetterName): Factor out the construction of the attribute setter function name.
(GenerateHeader): Determine that a class has read-write properties only if there is a read-write attribute that
is not static.
(GenerateAttributesHashTable): Skip static attributes in the object hash table. They will be added to the constructor
hash table.
(GenerateImplementation): Look for static attributes to add to the constructor hash table. Make a call to the static
function in the class.
* bindings/scripts/IDLParser.pm:
(ParseInterface): Update the processing because of the regex change.
* bindings/scripts/IDLStructure.pm: Update the attribute regex.
* bindings/scripts/test/JS/JSTestObj.cpp: Update test results.
* bindings/scripts/test/JS/JSTestObj.h: Update test results.
* bindings/scripts/test/TestObj.idl: Add test cases.

Identifier: 109370@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@122912 268f45cc-cd09-0410-ab3c-d52691b4dbfc