[Qt] QQuickWebView shouldn't recieve mouse events while dialogs are active https://bugs.webkit.org/show_bug.cgi?id=91634
Reviewed by Alexis Menard.

One side-effect of reimplementing childMouseEventFilter() in r122438 is that it can end up
bypassing the mouseArea mechanism used in QML dialogs and still forward some mouse events
over to the web process. We can rely on the same m_dialogActive mechanism already used for
touch events in there as well.

* UIProcess/API/qt/qquickwebview.cpp:
(QQuickWebView::childMouseEventFilter): ignore events and return early if a dialog is active.


Identifier: 109445@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@122987 268f45cc-cd09-0410-ab3c-d52691b4dbfc