IndexedDB: Add intVersion to chromium/public/WebIDBMetadata.h https://bugs.webkit.org/show_bug.cgi?id=91408
Reviewed by Adam Barth.

This is in support of the new upgradeneeded versioning api.
intVersion will eventually replace the WebString version member.

* public/WebIDBMetadata.h:
(WebKit::WebIDBMetadata::WebIDBMetadata):
The spec uses unsigned long long for version numbers but we use signed
long long so that we can use -1 as a sentinel. It indicates that a
database still uses a string version.


Identifier: 109517@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@123059 268f45cc-cd09-0410-ab3c-d52691b4dbfc