Crash in WebCore::StyleResolver::collectMatchingRulesForList https://bugs.webkit.org/show_bug.cgi?id=90803
Patch by Douglas Stockwell <dstockwell@google.com> on 2012-07-19
Reviewed by Andreas Kling.

Source/WebCore:

When a ProcessingInstruction was removed from the document the owner
was removed, but the style resolver was not guaranteed to be updated.
It was then possible for an inconsistent version of the stylesheet to
remain visible in the DOM. Fixed by removing an invalid condition and
mirroring the logic from StyleElement.

Test: fast/css/xml-stylesheet-removed.xhtml

* dom/ProcessingInstruction.cpp:
(WebCore::ProcessingInstruction::removedFrom): Mirror the logic from
StyleElement -- always update the style resolver.

LayoutTests:

* fast/css/xml-stylesheet-removed-expected.txt: Added.
* fast/css/xml-stylesheet-removed.xhtml: Added.

Identifier: 109586@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@123128 268f45cc-cd09-0410-ab3c-d52691b4dbfc