Keep overlap testing on for descendants of an animating element https://bugs.webkit.org/show_bug.cgi?id=91787
Reviewed by Beth Dakin.

Source/WebCore:

RenderLayerCompositor::computeCompositingRequirements() would turn off
overlap testing for descendants of a layer that has a 3D transform,
or has an animating transform. This could cause compositing changes
inside an animating element when transitions start and stop, which
is undesirable.

Instead, we can keep overlap testing on for descendants of that
animating element, but we have to turn it off for layers later
in z-order.

Also enhance the "IndirectCompositingReason" flags so that in
the compositing log output, we can distinguish between layers that
were composited as a result of overlap testing from those compositing
for stacking reasons.

Test: compositing/layer-creation/animation-overlap-with-children.html

* rendering/RenderLayer.h: New IndirectCompositingReason value to
distinguish overlap from stacking.
* rendering/RenderLayerCompositor.cpp:
(WebCore::RenderLayerCompositor::computeCompositingRequirements):
Use an IndirectCompositingReason enum rather than a bool to track whether
the layer starts as composited.
Remove the if (layer->has3DTransform() || isRunningAcceleratedTransformAnimation(layer->renderer()))
check that turned off overlap testing for descendants of this layer.
Use IndirectCompositingForStacking (somewhat arbitrarily) for reflection layers.
Turn overlap testing off for layers later in the z-order.
(WebCore::RenderLayerCompositor::requiresOwnBackingStore): Take IndirectCompositingForStacking into
account.
(WebCore::RenderLayerCompositor::reasonForCompositing): Improve the logging output.

LayoutTests:

Test with two layers inside an animating container; one should be composited
because of overlap, the other should not be.

* compositing/layer-creation/animation-overlap-with-children-expected.txt: Added.
* compositing/layer-creation/animation-overlap-with-children.html: Added.

Identifier: 109609@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@123151 268f45cc-cd09-0410-ab3c-d52691b4dbfc