Introduce Node::shadowHost() https://bugs.webkit.org/show_bug.cgi?id=91814
Reviewed by Hajime Morita.

Introduce Node::shadowHost(), which return the host element, or 0.
Node::shadowAncestorNode() is used to obtain a shadow host. However it
is confusing because it returns 'this' if this is not in a shadow tree.

Replaces one callsite of shadowAncestorNode() with shadowHost().

No behavior change.

* dom/Node.cpp:
(WebCore::Node::shadowHost): Added.
* dom/Node.h:
(Node): Declare shadowHost(), and add a comment to shadowAncestorNode().
* html/shadow/CalendarPickerElement.cpp:
(WebCore::CalendarPickerElement::hostInput):
Replace shadowAncestorNode() with shaodwHost().

Identifier: 109638@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@123180 268f45cc-cd09-0410-ab3c-d52691b4dbfc