[Gtk] Many tests revealed as passing after moving from Skipped to test_expectations.txt https://bugs.webkit.org/show_bug.cgi?id=85591
Unreviewed, managing a few more unexpected passes for the GTK port.

Remove a few expectations for tests that were passing, reclassify some failures as
architecture-specific and add PASS expectations for a few tests that pass even though
they test features that are currently not supported on the GTK port.

* platform/gtk/TestExpectations:


Identifier: 109682@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@123224 268f45cc-cd09-0410-ab3c-d52691b4dbfc