CodeGeneratorInspector.py is unnecessarily chatty https://bugs.webkit.org/show_bug.cgi?id=91758
Reviewed by Vsevolod Vlasov.

The code aded in r123091 included a "print" statement for each written
file. This makes for noisy build output, especially noticeable in the
Chromium ninja build (where build output takes up a single line of the
console).

If this print statement is generally useful, it should be
hidden behind a --verbose commandline option, as we do for the binding
generators.

* inspector/CodeGeneratorInspector.py:
(SmartOutput.close): Remove print statement.


Identifier: 109690@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@123233 268f45cc-cd09-0410-ab3c-d52691b4dbfc