<rdar://problem/11928576> SVG-as-image (constrained) intrinsic size calculation is wrong in vertical writing modes https://bugs.webkit.org/show_bug.cgi?id=91918
Reviewed by Anders Carlsson.

Source/WebCore:

Test: svg/as-image/svg-intrinsic-size-rectangular-vertical.html

* rendering/RenderReplaced.cpp:
(WebCore::RenderReplaced::computeAspectRatioInformationForRenderBox): When there is a content
renderer, it returns a physical intrinsic size and aspect ratio. The code was already
accounting for this by taking the reciprocal of the aspect ratio in the vertical case. Made
it also transpose the size itself, turning it from physical to logical. Moved this code
after setting m_intrinsicSize, since that member variable is always physical.

LayoutTests:

* svg/as-image/svg-intrinsic-size-rectangular-vertical-expected.html: Added.
* svg/as-image/svg-intrinsic-size-rectangular-vertical.html: Added.



Identifier: 109748@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@123294 268f45cc-cd09-0410-ab3c-d52691b4dbfc